refactor(etfs): mirror /etfs groups-index layout on /etfs/countries/<c>
Country root used to render a flat ETF listing; switch it to the
EtfGroupsIndex view (groups → categories → top ETFs) so it matches
the US `/etfs` page. Delete the now-redundant
`/etfs/countries/<c>/groups/page.tsx` — same data was reachable from
both, but US has no parallel `/etfs/groups` page, so non-US should
collapse the same way.

Knock-on changes:
- etfSectionIndexPath: collapse `groups` to base path for every country
  (was US-only); JSON-LD breadcrumb generators drop the redundant
  "Groups" intermediate crumb that pointed at the removed page.
- Add `/etfs-filtered/countries/[country]/page.tsx` mirror so the filter
  button still works from the country root (now a grid page); update
  `getEtfFilterDestination` to route `/etfs/countries/<c>`,
  `/etfs/countries/<c>/groups/<g>`, and the country-scoped asset-class /
  provider grids to that new mirror, lifting `group=<key>` as a filter
  param where applicable. Add `/etfs-filtered/countries/[^/]+` to the
  inline-filter pattern set so chip changes stay on the mirror.
- revalidateEtfGroupsIndexTag: CloudFront path is now the country root
  rather than `/etfs/countries/<c>/groups`.
- Drop the now-unused `generateEtfGroupsIndexMetadata` /
  `generateEtfGroupsIndexBreadcrumbJsonLd` exports.
