+++ insights-ui/src/app/etfs-filtered/countries/[country]/page.tsx
@@ -1,0 +1,41 @@
+import EtfPageLayout from '@/components/etfs/EtfPageLayout';
+import WithSuspenseEtfListingGrid from '@/components/etfs/WithSuspenseEtfListingGrid';
+import { fetchEtfListingData } from '@/utils/etf-data-utils';
+import { EtfSearchParams } from '@/utils/etf-filter-utils';
+import { resolveEtfCountryParam } from '@/utils/etf-country-route-utils';
+import { generateEtfCountryListingBreadcrumbJsonLd, generateEtfCountryListingMetadata } from '@/utils/etf-metadata-generators';
+import type { Metadata } from 'next';
+
+export const dynamic = 'force-dynamic';
+
+type PageProps = {
+  params: Promise<{ country: string }>;
+  searchParams: Promise<EtfSearchParams>;
+};
+
+export async function generateMetadata(props: { params: Promise<{ country: string }> }): Promise<Metadata> {
+  const { country } = await props.params;
+  const decoded = resolveEtfCountryParam(country, '/etfs');
+  return generateEtfCountryListingMetadata(decoded);
+}
+
+export default async function CountryEtfsFilteredPage({ params, searchParams: searchParamsPromise }: PageProps) {
+  const { country } = await params;
+  const decoded = resolveEtfCountryParam(country, '/etfs');
+
+  const searchParams = await searchParamsPromise;
+  const dataPromise = fetchEtfListingData(searchParams, decoded);
+
+  return (
+    <>
+      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(generateEtfCountryListingBreadcrumbJsonLd(decoded)) }} />
+      <EtfPageLayout
+        title={`${decoded} ETFs`}
+        description={`Explore ${decoded} exchange-traded funds with detailed financial metrics, expense ratios, dividend analysis, and AI-driven insights.`}
+        currentCountry={decoded}
+      >
+        <WithSuspenseEtfListingGrid dataPromise={dataPromise} />
+      </EtfPageLayout>
+    </>
+  );
+}
